+++ schematics/ng-update/data/input-names.ts
@@ -1,7 +1,7 @@
 import { InputNameUpgradeData, TargetVersion, VersionChanges } from '@angular/cdk/schematics';
 
 export const inputNames: VersionChanges<InputNameUpgradeData> = {
-    [TargetVersion.V15]: [
+    [TargetVersion.V16]: [
         {
             pr: '',
             changes: [
@@ -52,6 +52,13 @@ export const inputNames: VersionChanges<InputNameUpgradeData> = {
                     replaceWith: 'Desc',
                     limitedTo: {
                         attributes: ['thyActionMenuItemInfo', '[thyActionMenuItemInfo]']
+                    }
+                },
+                {
+                    replace: 'thyTitle',
+                    replaceWith: 'thyText',
+                    limitedTo: {
+                        elements: ['thy-action-menu-divider']
                     }
                 }
             ]