+++ schematics/ng-update/update-16/index.spec.ts
@@ -1,5 +1,29 @@
 import { HostTree, Tree } from '@angular-devkit/schematics';
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
+import { createTestWorkspaceFactory } from '../../testing';
+
+describe('ng-update v16 Schematic', () => {
+    let tree: Tree;
+    const schematicRunner = new SchematicTestRunner('migrations', require.resolve('../migration-collection.json'));
+
+    let workspaceTree: UnitTestTree;
+
+    beforeEach(async () => {
+        const factory = createTestWorkspaceFactory(schematicRunner);
+        await factory.create();
+        await factory.addApplication({ name: 'update-16-test' });
+
+        tree = factory.getTree();
+    });
+
+    it('should update to ng v16', async () => {
+        workspaceTree = await schematicRunner.runSchematic('migration-v16', undefined, tree);
+        const file = workspaceTree.get('package.json');
+        expect(file.content.toString()).toBeTruthy();
+        const packageJSON = JSON.parse(file.content.toString());
+        expect(packageJSON['dependencies']['@angular/core']).toContain('^16.');
+    });
+});
 
 describe('ng-update v16 Schematic', () => {
     let tree: Tree;