+++ schematics/ng-update/update-16/index.ts
@@ -1,9 +1,11 @@
 import { Rule } from '@angular-devkit/schematics';
 import { createMigrationSchematicRule, NullableDevkitMigration, TargetVersion } from '@angular/cdk/schematics';
-import { onMigrationComplete } from '../core/complete';
+import { onV16MigrationComplete } from '../core/complete';
 import { ruleUpgradeData } from '../data';
 import { AfterInsertElementMigration } from '../migrations/after-insert-element';
 import { BeforeInsertElementMigration } from '../migrations/before-insert-element';
+import { ActionMenuRemovalRule } from '../migrations/checks/action-menu-removal-rule';
+import { LabelRemovalRule } from '../migrations/checks/label-removal-rule';
 import { ClassNamesMigration } from '../migrations/class-names';
 import { EntryPointsMigration } from '../migrations/entry-points';
 import { InputNamesRemovalMigration } from '../migrations/input-names-removal';
@@ -17,9 +19,11 @@ const migrations: NullableDevkitMigration[] = [
     BeforeInsertElementMigration,
     AfterInsertElementMigration,
     ClassNamesMigration,
-    EntryPointsMigration
+    EntryPointsMigration,
+    ActionMenuRemovalRule,
+    LabelRemovalRule
 ];
 
 export default function main(): Rule {
-    return createMigrationSchematicRule(TargetVersion.V15, migrations, ruleUpgradeData, onMigrationComplete);
+    return createMigrationSchematicRule(TargetVersion.V16, migrations, ruleUpgradeData, onV16MigrationComplete);
 }