+++ src/core/behaviors/index.ts
@@ -6,3 +6,4 @@ export * from './loading-done';
 export * from './tabindex';
 export * from './unsubscribe';
 export * from './control-value-accessor';
+export * from './focus-control';