+++ src/dropdown/dropdown-menu.component.ts
@@ -1,16 +1,16 @@
+import { NgIf, NgTemplateOutlet } from '@angular/common';
 import {
+    ChangeDetectionStrategy,
+    Component,
     Directive,
     HostBinding,
     Input,
-    Component,
-    ViewEncapsulation,
-    ChangeDetectionStrategy,
+    TemplateRef,
     ViewChild,
-    TemplateRef
+    ViewEncapsulation
 } from '@angular/core';
+import { InputBoolean, InputCssPixel } from 'ngx-tethys/core';
 import { SafeAny } from 'ngx-tethys/types';
-import { InputBoolean } from 'ngx-tethys/core';
-import { NgIf } from '@angular/common';
 
 export type ThyDropdownMenuDividerType = 'default' | 'crossing' | '';
 
@@ -30,22 +30,26 @@ export class ThyDropdownAbstractMenu {
 @Component({
     selector: 'thy-dropdown-menu',
     template: `
-        <ng-container *ngIf="thyImmediateRender; then dropdownMenu"></ng-container>
+        <ng-container *ngIf="thyImmediateRender; then content"> </ng-container>
         <ng-template #dropdownMenu>
-            <div class="thy-dropdown-menu" [style.width.px]="width">
-                <ng-content></ng-content>
+            <div class="thy-dropdown-menu" [style.width]="thyWidth">
+                <ng-container *ngTemplateOutlet="content"></ng-container>
             </div>
+        </ng-template>
+        <ng-template #content>
+            <ng-content></ng-content>
         </ng-template>
     `,
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush,
-    host: {},
+    host: {
+        '[class.thy-dropdown-menu]': 'thyImmediateRender',
+        '[style.width]': "thyImmediateRender ? thyWidth : ''"
+    },
     standalone: true,
-    imports: [NgIf]
+    imports: [NgIf, NgTemplateOutlet]
 })
 export class ThyDropdownMenuComponent {
-    width: number;
-
     get template() {
         return this.templateRef;
     }
@@ -56,9 +60,9 @@ export class ThyDropdownMenuComponent {
      * 设置菜单宽度
      * @default 240px
      */
-    @Input() set thyWidth(value: number) {
-        this.width = value;
-    }
+    @Input()
+    @InputCssPixel()
+    thyWidth: number | string;
 
     /**
      * 是否直接渲染 dropdown-menu 中的元素